+++ CHANGES.md
@@ -169,4 +169,4 @@
   - bugfix: Fix (and document in python) the `to_enu` and `to_ned` functions in ITRFCoord
 
 ## Python 0.9.3
-  - bugfix: State transition matrix is transposed in python bindings for high-precision propagator
+  - bugfix: Fixed transposed state transition matrix in python bindings for high-precision propagator